refactoring the Admission Category wise Admission report Signed-off-by: Pubudu <pubudupiyankara.me@gmail.com>
